% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JScrap.R
\name{collectContentByKey}
\alias{collectContentByKey}
\title{Collect by keyword links from a web page}
\usage{
collectContentByKey(url, key)
}
\arguments{
\item{url}{url of the target web page}

\item{key}{keyword to search for}
}
\value{
list of bloc content containing that keyword
}
\description{
This function extract a webpage content by keyword, and return the blocs containing that keyword
}
