% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JScrap.R
\name{collectDataFromWebPage}
\alias{collectDataFromWebPage}
\title{Collect data from a web page}
\usage{
collectDataFromWebPage(excelpath, url)
}
\arguments{
\item{excelpath}{Path to the excel file}

\item{url}{url of the target web page11}
}
\value{
void
}
\description{
This function scraps data from a web page and save it into an excel sheet
Data such : Email, phones numbers, medias, links, social media links, and html tables
}
\details{
Example of usage :
collectDataFromWebPage(<path\file.xls>,"https://www.apec.fr/")
}
