% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JScrap.R
\name{collectDataFromWebSite}
\alias{collectDataFromWebSite}
\title{Collect data from the whole website}
\usage{
collectDataFromWebSite(url, excelpath, downloadpath)
}
\arguments{
\item{url}{url of the target web page}

\item{excelpath}{Path to the excel file}

\item{downloadpath}{path to download directory}
}
\value{
void
}
\description{
this function crawls the whole website for the requested data and save each scraped page into
an excel sheet, and download the availble media in a sepcific directory
}
