% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JScrap.R
\name{collectEmails}
\alias{collectEmails}
\title{Collect emails from a web page}
\usage{
collectEmails(url)
}
\arguments{
\item{url}{url of the target web page}
}
\value{
list of emails
}
\description{
this function extract all emails from a web page and return it as an R list
}
