% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studyCase.R
\name{collectFromYoutube}
\alias{collectFromYoutube}
\title{Collect youtube data}
\usage{
collectFromYoutube(type, id, rootPath)
}
\arguments{
\item{type}{element type to fetch ("activity","playlist","comment","subscription","caption")}

\item{id}{a video id or a channel id, video id for ("caption","comment") and channel id for the rest}

\item{rootPath}{path to save the generated results}
}
\value{
R dataframe representing the result node
}
\description{
This function collect data about a specific youtube element attached to videos and channels including activities, playlists, subscriptions, comments and video captions tracks
then parse it and save it into an R dataframe and an excel spredsheet file.
}
