% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JScrap.R
\name{collectMedias}
\alias{collectMedias}
\title{Collect media links from a web page}
\usage{
collectMedias(url, media)
}
\arguments{
\item{url}{url of the target web page}

\item{media}{media type could be : image,video,document,audio}
}
\value{
list of medias
}
\description{
this function exctract all medias such images, videos, audio, documents.. from a web page
}
