% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JScrap.R
\name{collectPhones}
\alias{collectPhones}
\title{Collect phone numbers from a web page}
\usage{
collectPhones(url)
}
\arguments{
\item{url}{url of the target web page}
}
\value{
list of phone numbers
}
\description{
this function extract all phone numbers from a web page and return it as an R list
}
