% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JScrap.R
\name{collectSocialLinks}
\alias{collectSocialLinks}
\title{Collect SocialLinks}
\usage{
collectSocialLinks(url)
}
\arguments{
\item{url}{url of the target web page}
}
\value{
list of social media links
}
\description{
this function extract all social medias links from a web page and return it as an R list
}
