% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JScrap.R
\name{collectWebSiteData}
\alias{collectWebSiteData}
\title{Collect data from a whole website}
\usage{
collectWebSiteData(url, excelpath, downloadpath)
}
\arguments{
\item{url}{url of the target web page}

\item{excelpath}{Path to the excel file}

\item{downloadpath}{path to the downloaded media files}
}
\value{
void
}
\description{
collectWebSiteData crawls a whole website for the requested data and save each scraped page into
an excel sheet, and download the availble media files in a sepcific directory
}
