% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JSocial.R
\name{countData}
\alias{countData}
\title{Count data rows in a json file}
\usage{
countData(path, item = "*", flatten = FALSE, rootarray)
}
\arguments{
\item{path}{Path to the json file}

\item{item}{columns to count}
}
\value{
number of rows in the target file
}
\description{
This function excutes an SQL select count(*) query over a json file and return data row count'
}
