% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JScrap.R
\name{downloadDocuments}
\alias{downloadDocuments}
\title{Download documents from a target page}
\usage{
downloadDocuments(path, url)
}
\arguments{
\item{path}{Path to the download directory}

\item{url}{url of the target web page}
}
\value{
void
}
\description{
This functions extract all documents files into a provided directory path
}
