% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JScrap.R
\name{downloadImages}
\alias{downloadImages}
\title{Download images from a target page}
\usage{
downloadImages(path, url)
}
\arguments{
\item{path}{Path to the download directory}

\item{url}{url of the target web page}
}
\value{
void
}
\description{
This function exctract all image files from a target webpage and download it into a provided directory path
Example of usage :
downloadImages(<path>,"https://www.wallpaper.com/art")
}
