% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JScrap.R
\name{downloadVideos}
\alias{downloadVideos}
\title{Download videos from a target page}
\usage{
downloadVideos(path, url)
}
\arguments{
\item{path}{Path to the download directory}

\item{url}{url of the target web page}
}
\value{
void
}
\description{
This functions extract all video files and download it into a provided directory path
}
