% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JSocial.R
\name{jsonLoadByListFromYoutube}
\alias{jsonLoadByListFromYoutube}
\title{Collect List of Data from youtube}
\usage{
jsonLoadByListFromYoutube(path, type, ids)
}
\arguments{
\item{path}{Path to save the json file}

\item{type}{type of element to retreive (channel, video, playlist)}

\item{ids}{array of ids}
}
\value{
void
}
\description{
this function calls youtube Google api to retreive json data of a iDs list and save into a file
}
