% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JSocial.R
\name{jsonLoadFromFacebook}
\alias{jsonLoadFromFacebook}
\title{Get facebook graph edge of a specific node}
\usage{
jsonLoadFromFacebook(path, node, edge, user = FALSE,
  params = rJava::.jarray(c("")))
}
\arguments{
\item{path}{Path to save the data}

\item{node}{Facebook graph node id, eg(Page id, User id, Post id)...}

\item{edge}{Graph edge name, eg(comment,posts,feed,likes) ...}

\item{user}{TRUE to use user token for the edge call, FALSE to use the App token}
}
\description{
this function calls facebook graph API endpoints to collect public data using nodes and edges
and save it into a json file
}
