% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JSocial.R
\name{jsonLoadFromYoutube}
\alias{jsonLoadFromYoutube}
\title{Collect Data from youtube}
\usage{
jsonLoadFromYoutube(path, type, id)
}
\arguments{
\item{path}{Path to save the json file}

\item{type}{type of element to retreive (channel, video, playlist)}

\item{id}{id of the element}
}
\value{
void
}
\description{
this function calls youtube Google api to retreive json data of a single given element and save into a file
}
