% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JSocial.R
\name{jsonLoadTopics}
\alias{jsonLoadTopics}
\title{Collect topics from Social medias such twitter and google+}
\usage{
jsonLoadTopics(path, query)
}
\arguments{
\item{path}{path of the json file to save}

\item{query}{Search query}
}
\value{
void
}
\description{
this function calls twitter and google+ api endpoints to retreive activities and tweets then save into json files
}
