% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JSocial.R
\name{jsonLoadUsers}
\alias{jsonLoadUsers}
\title{Collect data about users of facebook and google+}
\usage{
jsonLoadUsers(path, query)
}
\arguments{
\item{path}{Path to save file}

\item{query}{Search query}
}
\value{
void
}
\description{
this function collect and fetch data about facebook and google+ users and save it into a json file
}
