% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JSocial.R
\name{jsonSearchFomGooglePlus}
\alias{jsonSearchFomGooglePlus}
\title{Get google plus data}
\usage{
jsonSearchFomGooglePlus(path, query = "", lang = "en")
}
\arguments{
\item{path}{path to save the data}

\item{query}{text query to search for activities}

\item{lang}{language code of the activities ...}
}
\value{
void
}
\description{
this function fetch activities from google plus API by submitting a filtered search in save it to a json file
in the directory path provided
}
