% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JSocial.R
\name{jsonSearchFromFacebook}
\alias{jsonSearchFromFacebook}
\title{Search for facebook nodes with query}
\usage{
jsonSearchFromFacebook(path, node = "page", query)
}
\arguments{
\item{path}{Path to save the data}

\item{node}{Facebook node type : user, page, event, group}

\item{query}{Text query of the search request}
}
\description{
this function call the search endpoint for facebook nodes including users, pages, groups and events
by calling the facebook graph search endpoint, and save the result data into a json file
}
