% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JSocial.R
\name{jsonSearchFromYoutube}
\alias{jsonSearchFromYoutube}
\title{search Data from youtube}
\usage{
jsonSearchFromYoutube(path, type = "", query)
}
\arguments{
\item{path}{Path to save the json file}

\item{type}{type of element to retreive (channel, video, playlist)}

\item{query}{query of the request}
}
\value{
void
}
\description{
this function calls youtube search endpoint api to retreive json data of a given query or type and save into a file
}
