% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JSocial.R
\name{jsonSearchTweetsByFilter}
\alias{jsonSearchTweetsByFilter}
\title{Get Twitter data by filter}
\usage{
jsonSearchTweetsByFilter(path, lang = "en", exact = "", allword = "",
  hashtags = "", noneWords = "", oneOf = "", accounts = "",
  attitude = TRUE, question = FALSE)
}
\arguments{
\item{path}{directory path to save the json data file}

\item{lang}{language code of the tweets ...}

\item{attitude}{postive attitude : True, negative attitude : False}

\item{question}{question exist in tweet : True else false.}

\item{allwords}{filter tweets by all words in the string seperated by space}

\item{exactphrase}{filter tweets by exact string}

\item{hashtag}{filter tweets by hashtags in this string seperated by space ...}

\item{oneof}{filter tweets by one of the words in this string seperated by space ...}

\item{noneof}{filter tweets by excluding all the words in this string seperated by space ...}
}
\value{
void
}
\description{
this function target the twitter search endpoint of twitter API by filter and save the reponse data into a json file
in the directory path provided
}
