% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JSocial.R
\name{jsonSearchTweetsByQuery}
\alias{jsonSearchTweetsByQuery}
\title{Get Twitter data by text query}
\usage{
jsonSearchTweetsByQuery(path, query)
}
\arguments{
\item{path}{directory path to save the json data file}

\item{query}{fetch tweets by the provided text query}
}
\value{
void
}
\description{
this function target the twitter search endpoint of twitter API by text query and save the reponse data into a json file
in the directory path provided
}
