% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JScrap.R
\name{saveByKeyWord}
\alias{saveByKeyWord}
\title{Search content by key}
\usage{
saveByKeyWord(path, url, key)
}
\arguments{
\item{path}{Path to the excel file}

\item{url}{url of the target web page}

\item{key}{key to search for}
}
\value{
void
}
\description{
return html blocs by keyword search and save into excel sheet
}
