% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studyCase.R
\name{searchFacebookNodes}
\alias{searchFacebookNodes}
\title{Search facebook data}
\usage{
searchFacebookNodes(query, rootPath = "/home/dhaker/Desktop/Ghandi/")
}
\arguments{
\item{query}{text query to use for the search}

\item{rootPath}{path to save the generated results}
}
\description{
This function searchs facebook nodes including groups and events and pages,
then parse the results and save into an R dataframe and spredsheet excel file.
}
