% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studyCase.R
\name{searchFromFlickr}
\alias{searchFromFlickr}
\title{Search Flickr data}
\usage{
searchFromFlickr(query, rootPath)
}
\arguments{
\item{query}{text query to use for the search}

\item{rootPath}{path to save the generated results}
}
\description{
This function search for photos from Flickr based on a text query, then parse the result into an R dataframe and an excel spreedSheet
}
