% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studyCase.R
\name{searchFromGooglePlus}
\alias{searchFromGooglePlus}
\title{Search google plus data}
\usage{
searchFromGooglePlus(query, rootPath = "/home/dhaker/Desktop/Ghandi/")
}
\arguments{
\item{query}{text query to use for the search}

\item{rootPath}{path to save the generated results}
}
\description{
This function searchs google plus activities using a text query, then parse the result into an R dataframe and spreedSheet xls
}
