% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studyCase.R
\name{searchFromTumblr}
\alias{searchFromTumblr}
\title{Search Tumblr data}
\usage{
searchFromTumblr(query, rootPath)
}
\arguments{
\item{query}{text query to use for the search}

\item{rootPath}{path to save the generated results}
}
\description{
This function search for blog posts from Tumblr using a text query, then parse the result into an R dataframe and spreedSheet xls
}
