% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studyCase.R
\name{searchFromTwitter}
\alias{searchFromTwitter}
\title{Search Twitter data}
\usage{
searchFromTwitter(query, rootPath)
}
\arguments{
\item{query}{text query to use for the search}

\item{rootPath}{path to save the generated results}
}
\description{
This function searchs Tweets of the past seven days (week) from twitter using a text query, then parse the result into an R dataframe and also creates an excel spreedSheet
}
