% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studyCase.R
\name{searchFromYouTube}
\alias{searchFromYouTube}
\title{Search Youtube data}
\usage{
searchFromYouTube(query, rootPath = "/home/dhaker/Desktop/Ghandi/")
}
\arguments{
\item{query}{text query to use for the search}

\item{rootPath}{path to save the generated results}
}
\value{
R dataframe representing the search result
}
\description{
This function search for videos and channels from youtube using a text query, then parse the result into an R dataframe and spreedSheet xls
}
