% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JSocial.R
\name{selectData}
\alias{selectData}
\title{Filter data from json file}
\usage{
selectData(path, item = "id", where = "", file = "", append = FALSE,
  flatten = FALSE)
}
\arguments{
\item{path}{Path to the json file}

\item{item}{column to select}

\item{where}{an sql where clause.}

\item{file}{an optional argument to save the result into a provided file path}

\item{append}{TRUE to append file, FALSE to overwrite file}
}
\value{
array of result query
}
\description{
This function excutes an SQL select query over a json file and return filtered data'
}
