% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studyCase.R
\name{collectFacebookNodes}
\alias{collectFacebookNodes}
\title{Collect facebook data}
\usage{
collectFacebookNodes(nodeid, edge = "feed", user = FALSE,
  rootPath = "/home/dhaker/Desktop/Ghandi/")
}
\arguments{
\item{nodeid}{Id if the target facebook node}

\item{user}{FALSE to use an application based call, TRUE t o use Autenticated calls (used to fetch private groups and events)}

\item{rootPath}{path to save the generated results}
}
\description{
This function collect data from facebook (nodes/edge) couple such posts and feed in pages, groups and events,
then parse it and save into an R dataframe and spredsheet excel file.
}
