% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studyCase.R
\name{searchFromTwitter}
\alias{searchFromTwitter}
\title{Search Twitter data}
\usage{
searchFromTwitter(query, rootPath = "/home/dhaker/Desktop/Ghandi/")
}
\arguments{
\item{query}{text query to use for the search}

\item{rootPath}{path to save the generated results}
}
\description{
This function searchs Tweets of the last 7 days from twitter from using a text query, then parse the result into an R dataframe and spreedSheet xls
}
